The integrity checks are now less strict for non-source files
Failed integrity check (CRC32 hash comparasion) will terminate the installation only if the file was a .class file.
Otherwise, only a warning will be outputted into the console.
This way, when a few voicelines fail to download (or, more likely, download perfectly fine but their hashes weren't calculated properly), the update will still be able to be completed.
Also, the console debug output is much more verbose to help us with debugging.